requirements.txt
Flask==3.1.0
Flask-CORS==5.0.1
cos-python-sdk-v5>=1.9.30
PyJWT==2.10.1
cryptography==44.0.2
requests==2.32.3
python-dotenv==1.1.0
Pillow==11.1.0
Werkzeug==3.1.3
